Move URLSpanNoUnderline to native android package
